// src/frontend/src/declarations/backend.did.d.ts
export interface BlogPost {
  'id' : string,
  'title' : string,
  'content' : string,
  'published' : boolean,
  'createdAt' : Timestamp,
  'author' : string,
}
export type ExternalBlob = Uint8Array;
export interface Inquiry {
  'id' : string,
  'dob' : [] | [string],
  'tob' : [] | [string],
  'palmPhotos' : Array<[] | [ExternalBlob]>,
  'question' : string,
  'authorId' : [] | [Principal],
  'seedNumber' : [] | [bigint],
  'city' : [] | [string],
  'submittedAt' : Timestamp,
  'birthCountry' : [] | [string],
  'handPicture' : [] | [ExternalBlob],
  'state' : [] | [string],
  'visitorName' : string,
  'pastLifeNotes' : string,
  'serviceId' : bigint,
  'relationshipPerson2' : [] | [Person],
}
export interface Notice {
  'id' : string,
  'title' : string,
  'active' : boolean,
  'createdAt' : Timestamp,
  'message' : string,
}
export interface NumerologyUser {
  'username' : string,
  'passwordHash' : string,
  'sectionLevel' : bigint,
}
export interface Person {
  'dob' : [] | [string],
  'tob' : [] | [string],
  'name' : string,
}
export interface Service { 'id' : bigint, 'name' : string, 'price' : bigint }
export type Timestamp = bigint;
export interface UserProfile {
  'id' : Principal,
  'name' : string,
  'createdAt' : Timestamp,
  'email' : string,
}
export type UserRole = { 'admin' : null } |
  { 'user' : null } |
  { 'guest' : null };
export interface VisitorID {
  'service' : string,
  'expiresAt' : Timestamp,
  'username' : string,
  'password' : string,
  'createdAt' : Timestamp,
  'visitorName' : string,
}
export interface _CaffeineStorageCreateCertificateResult {
  'method' : string,
  'blob_hash' : string,
}
export interface _CaffeineStorageRefillInformation {
  'proposed_top_up_amount' : [] | [bigint],
}
export interface _CaffeineStorageRefillResult {
  'success' : [] | [boolean],
  'topped_up_amount' : [] | [bigint],
}
export interface _SERVICE {
  '_caffeineStorageBlobIsLive' : ActorMethod<[Uint8Array], boolean>,
  '_caffeineStorageBlobsToDelete' : ActorMethod<[], Array<Uint8Array>>,
  '_caffeineStorageConfirmBlobDeletion' : ActorMethod<
    [Array<Uint8Array>],
    undefined
  >,
  '_caffeineStorageCreateCertificate' : ActorMethod<
    [string],
    _CaffeineStorageCreateCertificateResult
  >,
  '_caffeineStorageRefillCashier' : ActorMethod<
    [[] | [_CaffeineStorageRefillInformation]],
    _CaffeineStorageRefillResult
  >,
  '_caffeineStorageUpdateGatewayPrincipals' : ActorMethod<[], undefined>,
  '_initializeAccessControlWithSecret' : ActorMethod<[string], undefined>,
  'assignCallerUserRole' : ActorMethod<[Principal, UserRole], undefined>,
  'createPost' : ActorMethod<[string, string, string], string>,
  'deleteInquiry' : ActorMethod<[string], undefined>,
  'deletePost' : ActorMethod<[string], undefined>,
  'getAllInquiries' : ActorMethod<[], Array<Inquiry>>,
  'getAllPosts' : ActorMethod<[], Array<BlogPost>>,
  'getAllPostsAdmin' : ActorMethod<[], Array<BlogPost>>,
  'getCallerUserProfile' : ActorMethod<[], [] | [UserProfile]>,
  'getCallerUserRole' : ActorMethod<[], UserRole>,
  'getServices' : ActorMethod<[], Array<Service>>,
  'getUserProfile' : ActorMethod<[Principal], [] | [UserProfile]>,
  'isCallerAdmin' : ActorMethod<[], boolean>,
  'noticeCreate' : ActorMethod<[string, string, string, string], string>,
  'noticeDelete' : ActorMethod<[string, string, string], undefined>,
  'noticeList' : ActorMethod<[], Array<Notice>>,
  'noticeListAll' : ActorMethod<[string, string], Array<Notice>>,
  'noticeToggleActive' : ActorMethod<[string, string, string], undefined>,
  'numerologyCreateUser' : ActorMethod<
    [string, string, string, string, bigint],
    undefined
  >,
  'numerologyDeleteUser' : ActorMethod<[string, string, string], undefined>,
  'numerologyListUsers' : ActorMethod<[string, string], Array<NumerologyUser>>,
  'numerologyLogin' : ActorMethod<[string, string], bigint>,
  'publishPost' : ActorMethod<[string], undefined>,
  'saveCallerUserProfile' : ActorMethod<[UserProfile], undefined>,
  'submitInquiry' : ActorMethod<[Inquiry], string>,
  'updatePost' : ActorMethod<[string, string, string, string], undefined>,
  'visitorCreateId' : ActorMethod<
    [string, string, string, string, string, string, bigint],
    undefined
  >,
  'visitorDeleteId' : ActorMethod<[string, string, string], undefined>,
  'visitorListIds' : ActorMethod<[string, string], Array<VisitorID>>,
  'visitorValidateId' : ActorMethod<[string, string, string], boolean>,
}
export declare const idlService: IDL.ServiceClass;
export declare const idlInitArgs: IDL.Type[];
export declare const idlFactory: IDL.InterfaceFactory;